+++ alinkjs/alink-plugins/AssetActionsPageSlotProvider.alink.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
     <h3 class="my-3">Actions:</h3>
-    <v-row class="ml-2" v-for="actionDef in assetLink.getRelevantActions({ asset })">
+    <v-row class="ml-2" v-for="actionDef in assetLink.getSlots({ type: 'asset-action', route: $route, asset })">
       <render-fn-wrapper class="mt-3" :key="actionDef.id" :render-fn="actionDef.componentFn"></render-fn-wrapper>
     </v-row>
   </div>
@@ -19,7 +19,9 @@ export default {
   inject: ['assetLink'],
   onLoad(handle) {
 
-    handle.definePageSlot('net.symbioquine.farmos_asset_link.slots.v0.actions', actionsSlot => {
+    handle.defineSlot('net.symbioquine.farmos_asset_link.slots.v0.actions', actionsSlot => {
+      actionsSlot.type('page-slot');
+
       actionsSlot.showIf(context => true);
 
       actionsSlot.componentFn((wrapper, h, context) => {